flask/pyproject.toml
[tool.poetry]
name = "flask_project"
version = "0.1.0"
description = ""
authors = ["Anastasiia <a.o.olifir@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
